Book Fly-In finally implemented
