+++ Capstone Project/Assets/PreFabs/Book.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 8317240110330064728}
   - component: {fileID: 3924577267478598763}
   - component: {fileID: 2155342038845974577}
+  - component: {fileID: 1299237384013595234}
   m_Layer: 0
   m_Name: Book
   m_TagString: Untagged
@@ -26,7 +27,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 8692562609108776154}
   serializedVersion: 2
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -0.53, y: 0.59, z: -0.21794309}
   m_LocalScale: {x: 0.5002, y: 0.2501, z: 0.5002}
   m_ConstrainProportionsScale: 0
@@ -110,3 +111,17 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
+--- !u!114 &1299237384013595234
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8692562609108776154}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 410b26000b6c3134c8aa98ec4c116c7f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  bookanimObj: {fileID: 8692562609108776154}
+  bSprite: {fileID: 3924577267478598763}
