+++ Capstone Project/Assets/Scripts/Player/BookBehaviour.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 410b26000b6c3134c8aa98ec4c116c7f
